backend/auth/requirements.txt
pydantic==2.5.0
psycopg2-binary==2.9.9
email-validator==2.1.0
